requirements.txt
# 🚀 FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 🗄️ Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# 🔐 Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# 📧 Email & SMS
sendgrid==6.11.0
twilio==8.10.0

# 📸 Image Management
imagekitio==3.2.0
Pillow==10.1.0

# 🔄 Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-dateutil==2.8.2

# 🧪 Testing (opcional)
pytest==7.4.3
httpx==0.25.2
